fixed tests and moved into correct folders. __init__.py files needed for correct import
